requirements.txt
fastapi
uvicorn
stripe

python-dotenv
